Moving logic not related to lineage back into metastore logic
